Adds playable deadites back. Rotting is now world time based, not tick based. - Rot timers adjusted. Cure rot for physicians removes deadite infection Zombie given a darkred overlay instead of a monochrome one. Mind now stores time of death. Decapitation should no longer cause issues. Observer adjustments

+++ code/modules/surgery/bodyparts/bodypart_dismemberment.dm
@@ -287,20 +287,14 @@
 		for(var/obj/item/worn_item in worn_items)
 			owner.dropItemToGround(worn_item, force = TRUE)
 
-//	owner.ghostize(0)
-//	if(brainmob)
-//		brainmob.ghostize(0)
-
+	if(isnull(owner))
+		return
 	qdel(owner.GetComponent(/datum/component/creamed)) //clean creampie overlay
-
 	name = "[owner.real_name]'s head"
+	owner.death()
+	owner.ghostize()
+
 	. = ..()
-	if(brainmob)
-		QDEL_NULL(brainmob)
-	var/obj/item/organ/brain/BR = locate(/obj/item/organ/brain) in contents
-	if(BR)
-		if(BR.brainmob)
-			QDEL_NULL(BR.brainmob)
 
 //Attach a limb to a human and drop any existing limb of that type.
 /obj/item/bodypart/proc/replace_limb(mob/living/carbon/C, special)
